requirements.txt
Django>=4.2.0,<5.0.0
graphene-django>=3.2.0,<4.0.0
django-filter>=23.0,<24.0
graphql-core>=3.2.0,<4.0.0
graphql-relay>=3.2.0,<4.0.0
